+++ easyadapter/RELEASE_NOTES.md
@@ -1,10 +1,12 @@
 [TOC]
 # Easy-adapter Release Notes
-
-## 1.0.0-alpha
+## 1.1.0-alpha
+##### Easyadapter
+* ANDDEP-646 - Improved EasyAdapter. In BaseItem added next and previous links to items, adapterPosition, position in ItemList.
+* ANDDEP-784 Added unbind method to BaseItemController, map for type ids
+## 1.0.0
 ##### Easyadapter
 * ANDDEP-270 Added support for asynchronous view inflate in ViewHolder
-* ANDDEP-784 Added unbind method to BaseItemController, map for type ids
 ## 0.3.0
 ##### Easyadapter
 * ANDDEP-200 - Removed ViewType randomization from ItemController